create way to access data from UI
While working on this issue, I realized it would be better for the view to just send commands directly to its `DataAccessor` instead of the App.  The App is used to set up the environment, not necessarily micro-manage everything.  In doing so, I also realized that the views may be better off being extensions of an abstract class (GUI is a View; Console is a View).