fix switch statement
original code would not compile.  After correcting syntax of switch statement cases from `-> {...;}` to `: ...; break;`, the code compiled and ran as expected.  Each window also matched the screenshots Josh provided.

I'm not sure if Josh wrote it that way or if it somehow got changed in a transfer over discord.