refactor SwingView to CardView
CardView is essentially identical to SwingView, but using CardLayout and each "page" is its own panel. The panels are changed when the relevant buttons are pressed.  each panel does not need to be recreated each time.

We may want to clear out all text fields before making the change.  It currently saves the stored data as the panel is not being recreated.