extract panel-specific methods to respective panels
Created a `void setSaveButton(JButton button, ActionListener actionListener)` for child panels to set the save button with the respective action.  The action was already being set by the respective panels.  If the button is not set by the panel, it does not exist.

A similar thing was done for the "Return to Results" button, but the action was maintained at the parent level -- this is just enabled or not.

These changes allowed removal of the `String viewType` parameter in the constructor, as well as the switch statement using the variable.

NOTE: Panel specific behavior should be defined within the respective panel, not in the abstract panel.  The abstract panel should not have any direct knowledge of the specific implementation of child classes, nor should it care.