remove unneeded files
These files were used for temporarily testing functions.  While they could have also served as the basis of fully implemented classes, it is beyond the scope of this course.  It is important to remember that full versions of these files could have been developed and used, especially if the customer desired a console-only application, for instance.