Create / Modify Record Field Fixes and Validation
Looking at the issues Akane brought up with the last name and date not saving for new records, I changed the code to generate the date based on today's date, disabling the field for user input. The last name was saving for me, so no changes were made there. I also added input validation for all fields of the create and modify records, throwing a unique exception for invalid input.